pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Starter template for making a new agent LangGraph."
authors = [
    { name = "FIXME", email = "FIXME@FIXME.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12,<4.0"
dependencies = [
    "azure-identity>=1.25.1,<2.0",
    "lagom>=2.7.7,<3.0",
    "langchain-openai>=1.1.7,<2.0",
    "langgraph-api>=0.5.27,<0.6.0",
    "langgraph>=1.0.0,<2.0",
    "pydantic-settings>=2.12.0,<3.0",
    "python-dotenv>=1.0.1,<2.0",
]


[project.optional-dependencies]
dev = [
    "mypy>=1.11.1,<2.0",
]

[dependency-groups]
cli = [
    # "pyright>=1.1.408,<2.0",  # TODO
    "ruff>=0.8.2,<0.9.0",
]
dev = [
    "anyio>=4.7.0,<5.0",
    "langgraph-cli[inmem]>=0.4.7,<0.5.0",
    "mypy>=1.13.0,<2.0",
    "pyhamcrest>=2.1.0,<3.0",
    "pytest>=8.3.5,<9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/agent"
]
